+++ plugins/field-angle/test/index.ts
@@ -1,0 +1,58 @@
+/**
+ * @license
+ * Copyright 2023 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * @fileoverview Angle field test playground.
+ */
+
+import * as Blockly from 'blockly';
+import {generateFieldTestBlocks, createPlayground} from '@blockly/dev-tools';
+import '../src/index';
+
+const toolbox = generateFieldTestBlocks('field_angle', [
+  {
+    label: 'Compass',
+    args: {
+      'mode': 'compass',
+      'value': 90,
+    },
+  },
+  {
+    label: 'Protractor',
+    args: {
+      'mode': 'protractor',
+      'value': 90,
+    },
+  },
+  {
+    label: 'Round',
+    args: {
+      'round': 1,
+    },
+  },
+]);
+
+/**
+ * Create a workspace.
+ * @param blocklyDiv The blockly container div.
+ * @param options The Blockly options.
+ * @returns The created workspace.
+ */
+function createWorkspace(blocklyDiv: HTMLElement,
+    options: Blockly.BlocklyOptions): Blockly.WorkspaceSvg {
+  const workspace = Blockly.inject(blocklyDiv, options);
+  return workspace;
+}
+
+document.addEventListener('DOMContentLoaded', function() {
+  const defaultOptions: Blockly.BlocklyOptions = {
+    toolbox,
+  };
+  const rootElement = document.getElementById('root');
+  if (rootElement) {
+    createPlayground(rootElement, createWorkspace, defaultOptions);
+  }
+});