+++ plugins/block-shareable-procedures/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.10](https://github.com/google/blockly-samples/compare/@blockly/block-shareable-procedures@1.2.9...@blockly/block-shareable-procedures@1.2.10) (2023-05-18)
+
+**Note:** Version bump only for package @blockly/block-shareable-procedures
+
+
+
+
+
+## [1.2.9](https://github.com/google/blockly-samples/compare/@blockly/block-shareable-procedures@1.2.8...@blockly/block-shareable-procedures@1.2.9) (2023-05-11)
+
+**Note:** Version bump only for package @blockly/block-shareable-procedures
+
+
+
+
+
 ## [1.2.8](https://github.com/google/blockly-samples/compare/@blockly/block-shareable-procedures@1.2.7...@blockly/block-shareable-procedures@1.2.8) (2023-05-04)
 
 **Note:** Version bump only for package @blockly/block-shareable-procedures