+++ plugins/cross-tab-copy-paste/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.19](https://github.com/google/blockly-samples/compare/@blockly/plugin-cross-tab-copy-paste@2.0.18...@blockly/plugin-cross-tab-copy-paste@2.0.19) (2023-05-18)
+
+**Note:** Version bump only for package @blockly/plugin-cross-tab-copy-paste
+
+
+
+
+
+## [2.0.18](https://github.com/google/blockly-samples/compare/@blockly/plugin-cross-tab-copy-paste@2.0.17...@blockly/plugin-cross-tab-copy-paste@2.0.18) (2023-05-11)
+
+**Note:** Version bump only for package @blockly/plugin-cross-tab-copy-paste
+
+
+
+
+
 ## [2.0.17](https://github.com/google/blockly-samples/compare/@blockly/plugin-cross-tab-copy-paste@2.0.16...@blockly/plugin-cross-tab-copy-paste@2.0.17) (2023-05-04)
 
 **Note:** Version bump only for package @blockly/plugin-cross-tab-copy-paste