+++ plugins/field-bitmap/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.18](https://github.com/google/blockly-samples/compare/@blockly/field-bitmap@2.0.17...@blockly/field-bitmap@2.0.18) (2023-05-18)
+
+**Note:** Version bump only for package @blockly/field-bitmap
+
+
+
+
+
+## [2.0.17](https://github.com/google/blockly-samples/compare/@blockly/field-bitmap@2.0.16...@blockly/field-bitmap@2.0.17) (2023-05-11)
+
+**Note:** Version bump only for package @blockly/field-bitmap
+
+
+
+
+
 ## [2.0.16](https://github.com/google/blockly-samples/compare/@blockly/field-bitmap@2.0.15...@blockly/field-bitmap@2.0.16) (2023-05-04)
 
 **Note:** Version bump only for package @blockly/field-bitmap