+++ plugins/renderer-inline-row-separators/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.13](https://github.com/google/blockly-samples/compare/@blockly/renderer-inline-row-separators@0.2.12...@blockly/renderer-inline-row-separators@0.2.13) (2023-05-18)
+
+**Note:** Version bump only for package @blockly/renderer-inline-row-separators
+
+
+
+
+
+## [0.2.12](https://github.com/google/blockly-samples/compare/@blockly/renderer-inline-row-separators@0.2.11...@blockly/renderer-inline-row-separators@0.2.12) (2023-05-11)
+
+**Note:** Version bump only for package @blockly/renderer-inline-row-separators
+
+
+
+
+
 ## [0.2.11](https://github.com/google/blockly-samples/compare/@blockly/renderer-inline-row-separators@0.2.10...@blockly/renderer-inline-row-separators@0.2.11) (2023-05-04)
 
 **Note:** Version bump only for package @blockly/renderer-inline-row-separators